separated db init from app start
