+++ backend/scripts/restrict_data.py
