+++ backend/scripts/zainscrape_classutil_courses.py
