+++ backend/src/models/db/index.js
@@ -1,23 +1,7 @@
 const { Connection, Request } = require('tedious')
 const { DB_CONFIG, MAX_CONNECTIONS } = require('./config')
-const { TABLE_COLUMNS, PRODUCTION } = require('../constants')
-const {
-    sqlTables,
-    sqlUniversity,
-    sqlFaculties,
-    sqlDegrees,
-    sqlSubjects,
-    sqlCourses,
-    sqlQuestions,
-    sqlReviews,
-    sqlComments,
-    sqlSessions,
-    sqlLikes,
-    sqlAdminUsers,
-    sqlUsers,
-    unswDataInitialised,
-    reviewTestDataInitialised
-} = require('./init_sql')
+const { TABLE_COLUMNS } = require('../constants')
+const { EventEmitter } = require('events')
 
 /**
  * Very slight abstraction over the direct sql queries.
@@ -25,10 +9,12 @@ const {
  * already filtered at this point.
  * @param {string} databaseName The name of the db if it needs to be passed in.
  */
-class DB {
+class DB extends EventEmitter {
     constructor() {
+        super()
         // Immediately start root connection process
         this.connections = [new Connection(DB_CONFIG)]
+        this.ready = false
 
         // Setup a pool of other connections
         for (let i = 1; i < MAX_CONNECTIONS; i++) {
@@ -39,58 +25,13 @@ class DB {
                     throw new Error('Couldn\'t connect to DB')
                 }
                 this.connections.push(connection)
+                // emit ready event once only when a connection becomes available
+                if (!this.ready) {
+                    this.ready = true
+                    this.emit('ready')
+                }
             })
         }
-    }
-
-    async init() {
-        return new Promise((resolve, reject) => {
-            // Database initialisation benchmarking
-            const timeList = [Date.now() / 1000]
-
-            // Do the initialisation
-            const [connection] = this.connections
-
-            connection.on('connect', (err) => {
-                if (err) reject(err)
-
-                // Create the database and initialise data with no dependencies.
-                const request = new Request(sqlTables(), async (err) => {
-                    if (err) reject(err)
-
-                    // Initialise with UNSW data
-                    if (!await unswDataInitialised(connection)) {
-                        await sqlUniversity(connection)
-                        await sqlFaculties(connection)
-                        await sqlDegrees(connection)
-                        await sqlSubjects(connection)
-                        await sqlCourses(connection)
-                        await sqlSessions(connection)
-                    }
-
-                    // user and question tables get dropped on init, so we must always re-add them
-                    await sqlAdminUsers(connection)
-                    await sqlQuestions(connection)
-
-                    // Initialise testing data
-                    // uses review table to check if initialization has happened
-                    if (!PRODUCTION && !await reviewTestDataInitialised(connection)) {
-                        await sqlReviews(connection)
-                        await sqlComments(connection)
-                        await sqlLikes(connection)
-                        await sqlUsers(connection)
-                    }
-
-                    // Log completion time
-                    timeList.push(Date.now() / 1000)
-                    console.log(`Done creating database! (${((timeList[1] - timeList[0])).toFixed(3)})`)
-
-                    // Done
-                    resolve()
-                })
-                connection.execSql(request)
-            })
-        })
     }
 
     delete() {
@@ -137,5 +78,7 @@ class DB {
         this.connections.forEach((connection) => connection.close())
     }
 }
+
+//util.inherits(DB, EventEmitter)
 
 module.exports = new DB()
